+++ fc2cmadb_crawler/copy_non_media_files.py
@@ -1,5 +1,0 @@
-from .copy_non_media import main
-
-
-if __name__ == "__main__":
-    raise SystemExit(main())